// components/types.ts
export type CarConfig = {
    bodyColor: string
    wheelColor: string
    windowTint: number
    metalness: number
    roughness: number
    spoilerType: "none" | "sport" | "racing"
    rimStyle: "standard" | "sport" | "chrome"
    headlightColor: string
    underglow: boolean
    underglowColor: string
}